pyproject.toml
[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[project]
name = "persona-stack"
version = "0.1.0"
description = "Local SDXL persona generation stack (ComfyUI automation)"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Imaginarium" }]
keywords = ["comfyui", "sdxl", "image-generation", "workflow"]
requires-python = ">=3.11"
dependencies = [
  "huggingface-hub>=0.20.0",
  "pillow>=10.0.0",
  "requests>=2.31.0",
  "typer>=0.12.0",
  "websocket-client>=1.7.0",
]

[project.scripts]
persona-stack = "persona_stack.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/persona_stack"]
